backend/requirements.txt
# Web Framework and API
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.1
pydantic-settings==2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# Blockchain Integration
web3==6.11.3
eth-account==0.9.0

# CORS and Middleware
starlette==0.27.0

# Testing
pytest==7.4.3
httpx==0.25.1

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0